refactor: implement nestjs module model
