+++ docker-compose.yml
@@ -13,7 +13,6 @@ services:
         DOC_PORT: ${DOC_PORT}
     networks:
       users-bridge:
-        ipv4_address: ${NETWORK_APP_IP}
     depends_on:
       postgres:
         condition: service_healthy
@@ -23,7 +22,7 @@ services:
       POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
       PORT: ${PORT}
       DOC_PORT: ${DOC_PORT}
-      POSTGRES_HOST: ${DB_IP}
+      #POSTGRES_HOST: ${DB_IP}
       POSTGRES_PORT: ${POSTGRES_PORT}
       SSH_PRIVKEY: ${SSH_PRIVKEY}
       SSH_CERT: ${SSH_CERT}
@@ -37,7 +36,7 @@ services:
     ports:
       - ${PORT}:${PORT}
       - ${DOC_PORT}:${DOC_PORT}
-    restart: unless-stopped
+    restart: on-failure
   postgres:
     build:
       dockerfile: ${BASE_DOCKERFILE}
@@ -49,7 +48,6 @@ services:
       - "logging_collector=on"
     networks:
       users-bridge:
-        ipv4_address: ${DB_IP}
     environment:
       POSTGRES_DB: ${POSTGRES_DATABASE}
       POSTGRES_USER: ${POSTGRES_USER}
@@ -73,10 +71,5 @@ services:
 networks:
   users-bridge:
     driver: bridge
-    ipam:
-      driver: default
-      config:
-        - subnet: ${NETWORK_SUBNET}
-          gateway: ${NETWORK_GATEWAY}
 volumes:
   pg_data:
