+++ src/auth/auth.module.ts
@@ -1,20 +1,16 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
-import { TypeOrmModule } from '@nestjs/typeorm';
 import { LoggerModule } from 'src/logging/log.module';
-import { appDataSource } from 'src/utils/db/dataSource';
-import { UserEntity } from 'src/routes/user/user.entity';
 import { AuthController } from './auth.controller';
 import { AuthService } from './authService';
 import { JwtStrategy } from './jwt.strategy';
 import { RefreshTokenStrategy } from './refresh.strategy';
-import { UserService } from 'src/routes/user/user.service';
+import { UserModule } from 'src/user/user.module';
 
 @Module({
   imports: [
-    TypeOrmModule.forRoot(appDataSource.options),
-    TypeOrmModule.forFeature([UserEntity]),
+    UserModule,
     PassportModule,
     LoggerModule,
     JwtModule.register({
@@ -24,7 +20,7 @@ import { UserService } from 'src/routes/user/user.service';
       },
     }),
   ],
-  providers: [UserService, AuthService, JwtStrategy, RefreshTokenStrategy],
+  providers: [AuthService, JwtStrategy, RefreshTokenStrategy],
   controllers: [AuthController],
   exports: [AuthService],
 })
