+++ src/auth/authService.ts
@@ -2,7 +2,7 @@ import { ForbiddenException, Injectable } from '@nestjs/common';
 import { UnauthorizedException } from '@nestjs/common/exceptions';
 import { JwtService } from '@nestjs/jwt';
 import { JsonWebTokenError } from 'jsonwebtoken';
-import { UserService } from 'src/routes/user/user.service';
+import { UserService } from 'src/user/user.service';
 import { compare } from 'src/utils/hash';
 
 @Injectable()
