+++ src/musicLibrary/album/album.service.ts
@@ -1,6 +1,9 @@
 import { forwardRef, Inject, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { FavoritesDto, FavoritesType } from 'src/routes/favs/favorites.dto';
+import {
+  FavoritesDto,
+  FavoritesType,
+} from 'src/musicLibrary/favs/favorites.dto';
 import { Repository } from 'typeorm';
 import { EntityService } from '../entity.service';
 import { FavoritesService } from '../favs/favorites.service';
