+++ src/musicLibrary/artist/artist.entity.ts
