+++ src/musicLibrary/track/track.service.ts
@@ -1,7 +1,10 @@
 import { forwardRef, Inject, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { FavoritesDto, FavoritesType } from 'src/routes/favs/favorites.dto';
-import { TrackDto } from 'src/routes/track/track.dto';
+import {
+  FavoritesDto,
+  FavoritesType,
+} from 'src/musicLibrary/favs/favorites.dto';
+import { TrackDto } from 'src/musicLibrary/track/track.dto';
 import { Repository } from 'typeorm';
 import { EntityService } from '../entity.service';
 import { FavoritesService } from '../favs/favorites.service';
