+++ src/user/user.module.ts
@@ -1,0 +1,17 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { appDataSource } from 'src/utils/db/dataSource';
+import { UserController } from './user.controller';
+import { UserEntity } from './user.entity';
+import { UserService } from './user.service';
+
+@Module({
+  imports: [
+    TypeOrmModule.forRoot(appDataSource.options),
+    TypeOrmModule.forFeature([UserEntity]),
+  ],
+  controllers: [UserController],
+  providers: [UserService],
+  exports: [UserService],
+})
+export class UserModule {}
